requirements.txt
# python version: 3.13.2
langchain==1.0.8
langgraph==1.0.3
deepagents==0.2.7

langchain-community==0.4.1
langchain-deepseek==1.0.1
langchain-openai==1.0.3
langchain-ollama==1.0.0
langchain-anthropic==1.1.0
langchain-groq==1.0.1
langchain-google-genai==3.1.0
langchain-mistralai==1.0.1
langchain-huggingface==1.0.1
langchain-xai==1.0.0

langchain-mcp-adapters==0.1.13
mcp==1.22.0
mcp[cli]==1.22.0
tavily-python==0.7.13
langgraph-checkpoint-sqlite==3.0.0
aiosqlite==0.21.0

rich==14.2.0
prompt_toolkit==3.0.52

Nuitka==2.8.6
pyinstaller==6.16.0
pytest==9.0.1